refactor: rename I18n component to MyI18n and update references in stories to fix linting errors
