+++ examples/showcase/components/MyI18n.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/vue3'
 
-import MyComponent from './I18n.vue'
+import MyComponent from './MyI18n.vue'
 
 // More on how to set up stories at: https://storybook.js.org/docs/vue/writing-stories/introduction
